fix: fix Editor & translate
