+++ renderer/public/locales/en/common.json
@@ -21,6 +21,7 @@
   "divider": "Selecting verses for translation",
   "Download": "Download",
   "editor": "Editor",
+  "emptyAlphabet": "Your dictionary is empty",
   "EmptyFileContent": "Empty file content",
   "EmptyTitle": "Empty Title",
   "EN": "EN",
@@ -42,6 +43,7 @@
   "NoFileSelected": "No file selected",
   "NoMatches": "No matches",
   "NotSelected": "Not selected",
+  "NoWords": "Your dictionary does not contain a single word. Click \"+\" to add the first word",
   "Ok": "Ok",
   "Open": "Open",
   "OriginalRecording": "Retelling in the original language:",
