+++ renderer/public/locales/es/common.json
@@ -21,6 +21,7 @@
   "divider": "Seleccionar versos para traducir",
   "Download": "Descargar",
   "editor": "Editor",
+  "emptyAlphabet": "Tu diccionario está vacío",
   "EmptyFileContent": "Contenido del archivo vacío",
   "EmptyTitle": "Título vacío",
   "EN": "EN",
@@ -42,6 +43,7 @@
   "NoFileSelected": "Ningún archivo seleccionado",
   "NoMatches": "No hay coincidencias",
   "NotSelected": "No seleccionado",
+  "NoWords": "Tu diccionario no contiene una sola palabra. Presione \"+\" para agregar la primera palabra",
   "Ok": "Ok",
   "Open": "Abierto",
   "OriginalRecording": "Volver a contar en la lengua original:",
