+++ renderer/public/locales/ru/common.json
@@ -21,6 +21,7 @@
   "divider": "Выбор стихов для перевода",
   "Download": "Скачать",
   "editor": "Редактор",
+  "emptyAlphabet": "Ваш словарь пуст",
   "EmptyFileContent": "Отсутствует содержимое файла",
   "EmptyTitle": "Пустой заголовок",
   "EN": "EN",
@@ -42,6 +43,7 @@
   "NoFileSelected": "Не выбран файл",
   "NoMatches": "Нет совпадений",
   "NotSelected": "Не выбран",
+  "NoWords": "Ваш словарь не содержит ни одного слова. Нажмите \"+\" чтобы добавить первое слово",
   "Ok": "Понятно",
   "Open": "Открыть",
   "OriginalRecording": "Пересказ на исходном языке:",
